Use ChosenInlineResultHandler to populate messages
The workflow is significantly modified with this
commit. The InlineQueryResultArticle objects don't
contain the entire dictionary text of a word
anymore. When a user selects a query result, a
placeholder message will be sent which whill be
filled afterwards.

This allows to send more InlineQueryResultArticle
objects without filling the message length limit.
So, we can send all entries that match the query.
This leads to a MUCH better user experience.